requirements.txt
alembic==1.0.8
argon2-cffi==19.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
certifi==2020.11.8
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.7
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.10
importlib-metadata==0.15
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.0.9
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.2
numpy-stl==2.10.1
passlib==1.7.1
Pillow==6.2.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pytest==4.5.0
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
python-magic==0.4.15
python-openctm==1.0.11
python-utils==2.3.0
requests==2.25.0
six==1.12.0
SQLAlchemy==1.3.11
trimesh==3.9.19
urllib3==1.26.2
wcwidth==0.1.7
Werkzeug==0.15.6
WTForms==2.2.1
zipp==0.5.1
